Feat/file upload

+++ src/test/java/de/mherrmann/famkidmem/backend/TestUtils.java
@@ -51,6 +51,9 @@ public class TestUtils {
 
     public void deleteTestFiles() {
         File directory = new File("./files");
+        if(!directory.exists()){
+            return;
+        }
         for (File file : directory.listFiles()) {
             file.delete();
         }
@@ -172,7 +175,7 @@ public class TestUtils {
     }
 
     public void createTestFile(String filename) throws IOException {
-        new File("./files").mkdir();
+        createTestFilesDirectory();
         new File("./files/"+filename).createNewFile();
         FileWriter fileWriter = new FileWriter("./files/"+filename);
         fileWriter.write(filename);
@@ -186,8 +189,12 @@ public class TestUtils {
         fileWriter.close();
     }
 
-    public void deleteAuthTokenHashFile() throws IOException {
+    public void deleteAuthTokenHashFile() {
         new File("./ccms_auth_token_hash").delete();
+    }
+
+    public void createTestFilesDirectory(){
+        new File("./files").mkdir();
     }
 
 }
